# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/appointments": {
      "post": {
        "operationId": "create",
        "summary": "Create a new appointment",
        "description": "Create a new appointment for the authenticated user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid appointment data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pet, clinic, or user not found"
          },
          "409": {
            "description": "Appointment scheduling conflict"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all appointments",
        "description": "Retrieve all appointments with optional filtering and pagination",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by appointment status",
            "schema": {
              "enum": [
                "pending",
                "confirmed",
                "in_progress",
                "completed",
                "cancelled",
                "no_show",
                "rescheduled",
                "waiting"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by appointment type",
            "schema": {
              "enum": [
                "consultation",
                "vaccination",
                "surgery",
                "follow_up",
                "emergency",
                "wellness_exam",
                "dental_cleaning",
                "laboratory_test",
                "imaging",
                "therapy",
                "grooming",
                "behavioral_training",
                "nutrition_consultation",
                "physical_therapy",
                "specialist_consultation"
              ],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filter by appointment priority",
            "schema": {
              "enum": [
                "low",
                "normal",
                "high",
                "urgent",
                "emergency"
              ],
              "type": "string"
            }
          },
          {
            "name": "clinic_id",
            "required": false,
            "in": "query",
            "description": "Filter by clinic ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "staff_id",
            "required": false,
            "in": "query",
            "description": "Filter by staff ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pet_id",
            "required": false,
            "in": "query",
            "description": "Filter by pet ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner_id",
            "required": false,
            "in": "query",
            "description": "Filter by owner ID (admin only)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_from",
            "required": false,
            "in": "query",
            "description": "Filter by start date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_to",
            "required": false,
            "in": "query",
            "description": "Filter by end date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_telemedicine",
            "required": false,
            "in": "query",
            "description": "Filter by telemedicine appointments",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "is_home_visit",
            "required": false,
            "in": "query",
            "description": "Filter by home visit appointments",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in pet name, owner name, or notes",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointments retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointments/my-appointments": {
      "get": {
        "operationId": "findMyAppointments",
        "summary": "Get current user appointments",
        "description": "Retrieve all appointments owned by the authenticated user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User appointments retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointments/pet/{petId}": {
      "get": {
        "operationId": "findByPet",
        "summary": "Get appointments by pet",
        "description": "Retrieve all appointments for a specific pet",
        "parameters": [
          {
            "name": "petId",
            "required": true,
            "in": "path",
            "description": "Pet ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet appointments retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointments/clinic/{clinicId}": {
      "get": {
        "operationId": "findByClinic",
        "summary": "Get appointments by clinic",
        "description": "Retrieve all appointments for a specific clinic",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "description": "Clinic ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Filter by specific date (ISO string)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clinic appointments retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointments/staff/{staffId}": {
      "get": {
        "operationId": "findByStaff",
        "summary": "Get appointments by staff member",
        "description": "Retrieve all appointments for a specific staff member",
        "parameters": [
          {
            "name": "staffId",
            "required": true,
            "in": "path",
            "description": "Staff ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Filter by specific date (ISO string)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Staff appointments retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointments/stats": {
      "get": {
        "operationId": "getAppointmentStats",
        "summary": "Get appointment statistics",
        "description": "Retrieve comprehensive statistics about all appointments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Appointment statistics retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointments/available-slots/{clinicId}": {
      "get": {
        "operationId": "getAvailableTimeSlots",
        "summary": "Get available time slots",
        "description": "Get available time slots for a specific clinic and date",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "description": "Clinic ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Date to check (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "duration",
            "required": false,
            "in": "query",
            "description": "Duration in minutes (default: 30)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available time slots retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointments/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get appointment by ID",
        "description": "Retrieve a specific appointment by its ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Appointment not found"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "update",
        "summary": "Update appointment",
        "description": "Update a specific appointment by its ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid update data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Appointment not found"
          },
          "409": {
            "description": "Appointment scheduling conflict"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Cancel appointment",
        "description": "Cancel a specific appointment by its ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment cancelled successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Appointment not found"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointments/{id}/status": {
      "patch": {
        "operationId": "updateStatus",
        "summary": "Update appointment status",
        "description": "Update the status of a specific appointment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "description": "New appointment status",
            "schema": {
              "enum": [
                "pending",
                "confirmed",
                "in_progress",
                "completed",
                "cancelled",
                "no_show",
                "rescheduled",
                "waiting"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid status"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Appointment not found"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointments/{id}/reschedule": {
      "patch": {
        "operationId": "rescheduleAppointment",
        "summary": "Reschedule appointment",
        "description": "Reschedule a specific appointment to a new date/time",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "new_date",
            "required": true,
            "in": "query",
            "description": "New appointment date (ISO string)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment rescheduled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid new date"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Appointment not found"
          },
          "409": {
            "description": "New time conflicts with existing appointment"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Clinic Management API",
    "description": "AI-Powered Pet Clinic Management Platform Backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}